chore: split the vault handler test, taking 25 tests out of the mock blast radius
- What: the four pure describes (parseGitStatus, mapToVaultPath, mapFromVaultPath,
  ensureFrontmatterProject) move to `handler-paths.test.ts` and now import from
  their SOURCE modules — `../git.ts`, `../path-mapping.ts` — instead of through
  `../handler.ts`. handler.test.ts 433 → 250; allow-list entry removed.

- Why beyond size: `handler.test.ts` registers process-wide `mock.module()` for
  `../../db/index.ts` and `../discovery.ts`. CLAUDE.md documents that those leak
  into sibling files and fail two unrelated `syncVault lock` tests whenever
  `--isolate` is omitted. The pure functions never needed those mocks — they were
  only reached through handler.ts, which does.

- It does NOT fix the leak, and the comment says so. Measured without `--isolate`,
  before and after: the same two `syncVault lock` tests fail either way. What
  changed is that 25 tests no longer participate.

Test count verified identical: 55 before, 55 after, 97 expect() calls both ways.

That check mattered — my first before/after showed 80 vs 55 and looked like 25 lost
tests. `git stash push` without `-u` does not stash UNTRACKED files, so the new file
stayed on disk during the "before" run and its 25 tests were counted twice. The
measurement was wrong, not the split.

Gate: `bunx tsc --noEmit` exit 0 · `bun test --isolate src/vault/ tests/vault/
tests/build/` 134 pass.

Allow-list: 22 at #2833's filing → **13**.

Refs #2833
